Moves types over into the models folder. Adds documentation for types using Input
