+++ src/components/design/forms/Date.tsx
@@ -1,16 +1,22 @@
 import React from "react";
-import { Input, InputProps } from "./Input";
+import { Input } from "./Input";
+import { CoreInputProps } from "../../../models/design/form";
 
 /**
  * An input set up for handling dates. Formats the Input function to work properly
  * @param props.aria-label The usability label
  * @param props.disabled If an input is disabled or not
+ * @param props.id The input id
+ * @param props.label The label of the inputs
  * @param props.name The name of the input
  * @param props.placeholder A placeholder value for when the input is empty
- * @param props.rows The number of rows (multiline must be true)
- * @param props.as Controls what the input is seen as
  * @param props.size The bootstrap size of the input. 
+ * @param props.xs Width for extra small screens
+ * @param props.sm Width for small screens
+ * @param props.md Width for extra medium screens
+ * @param props.lg Width for large screens
+ * @param props.xl Width for extra large screens
  */
-export function Date(props: InputProps) {
+export function Date(props: CoreInputProps) {
   return <Input {...props} type="date"/>;
 }
