+++ src/components/design/forms/Input.tsx
@@ -1,29 +1,37 @@
 /* eslint-disable @typescript-eslint/no-use-before-define */
 import React from "react";
 import { Form } from "react-bootstrap"
-import { FieldProps } from "./types";
 
 import { useField } from "formik";
+import { CoreInputProps } from "../../../models/design/form";
 
+// The type describing valid inputs for Input's as field
 type As = "input" | "textarea";
 
-export interface InputProps extends FieldProps {
-  // Default input properties
-  "aria-label"?: string; // The usability label
-  disabled?: boolean; // If an input is disabled or not
-  name: string; // The name of the input
-  placeholder?: string; // A placeholder value for when the input is empty
-  rows?: number; // The number of rows (multiline must be true)
+interface InputProps extends CoreInputProps {
+  as?: As; // Controls what the input is seen as (input or textarea)
+  rows?: number; // The number of rows (as must be textarea)
   type?: string; // The type of input (text, password, datetime, etc.)
-
-  // Bootstrap Properties
-  as?: As; // Controls what the input is seen as
-  size?: "sm" | "lg"; // The bootstrap size of the input
 }
+
 
 /**
  * Renders out a grid-based input block
- * @param props see IInput
+ * @param props.as Controls what the input is seen as
+ * @param props.aria-label The usability label
+ * @param props.disabled If an input is disabled or not
+ * @param props.id The input id
+ * @param props.label The label of the inputs
+ * @param props.name The name of the input
+ * @param props.placeholder A placeholder value for when the input is empty
+ * @param props.rows The number of rows (`as` must be textarea)
+ * @param props.type The type of the 
+ * @param props.size The bootstrap size of the input. 
+ * @param props.xs Width for extra small screens
+ * @param props.sm Width for small screens
+ * @param props.md Width for extra medium screens
+ * @param props.lg Width for large screens
+ * @param props.xl Width for extra large screens
  */
 export function Input(props: InputProps) {
   const [field] = useField(props);
